Update bb.org_4gb_stable.sh
Not sure if I should hard code this username (buildbot), but it'd be nice if the image wasn't owned by root in the end. Unsure about all the moving.

TODO: If suitable, I might add an 'rsync' or 'scp' directly of the file to a location on beagleboard.org where the file can be served. Other option is to simply use the buildbot file upload script, which could be fine if we are talking about a single image and only test versions.